Fix in resetEmbedded when textTracks is null

+++ src/streaming/text/TextSourceBuffer.js
@@ -170,8 +170,9 @@ function TextSourceBuffer() {
 
     function resetEmbedded() {
         eventBus.off(Events.VIDEO_CHUNK_RECEIVED, onVideoChunkReceived, this);
-
-        textTracks.deleteAllTextTracks();
+        if (textTracks !== null) {
+            textTracks.deleteAllTextTracks();
+        }
         embeddedInitialized = false;
         embeddedTracks = [];
         embeddedCea608FieldParsers = [null, null];
